app/requirements.txt
Flask
psycopg2-binary
prometheus-client==0.20.0
